Worked on pathfinding for variable terrain cost.

+++ src/terrain/Inter.java
@@ -7,7 +7,7 @@ public interface Inter {
 	boolean drive();
 	boolean swim();
 	boolean fly();
-	int speed();
+	double speed();
 	int defense();
 	int x();
 	int y();
